+++ tests/Utils/Models/User.php
@@ -8,6 +8,8 @@ use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\MorphOne;
 use Illuminate\Foundation\Auth\User as Authenticatable;
+use Tests\DBTestCase;
+use Tests\Integration\Execution\DataLoader\RelationBatchLoaderTest;
 
 /**
  * Primary key.
@@ -30,6 +32,7 @@ use Illuminate\Foundation\Auth\User as Authenticatable;
  * @property string|null $person_type
  *
  * Relations
+ * @property-read \Illuminate\Database\Eloquent\Collection<\Tests\Utils\Models\AlternateConnection> $alternateConnections
  * @property-read \Tests\Utils\Models\Company|null $company
  * @property-read \Tests\Utils\Models\Image|null $image
  * @property-read \Illuminate\Database\Eloquent\Collection<\Tests\Utils\Models\Post> $posts
@@ -40,6 +43,18 @@ use Illuminate\Foundation\Auth\User as Authenticatable;
  */
 class User extends Authenticatable
 {
+    /**
+     * Ensure that this is functionally equivalent to leaving this as null.
+     *
+     * @see RelationBatchLoaderTest::testDoesNotBatchloadRelationsWithDifferentDatabaseConnections()
+     */
+    protected $connection = DBTestCase::DEFAULT_CONNECTION;
+
+    public function alternateConnections(): HasMany
+    {
+        return $this->hasMany(AlternateConnection::class);
+    }
+
     public function company(): BelongsTo
     {
         return $this->belongsTo(Company::class);